+++ scripts/install.ps1
@@ -10,6 +10,7 @@ param(
 
 $Owner   = "Data-Corruption"
 $Repo    = "sprout"
+
 $AppName = "sprout"
 $Service = $true